Convert percentages stored as a string to numeric
Further to https://github.com/PHPOffice/PhpSpreadsheet/issues/3155, I noticed that percentages stored as strings in Excel were still being used in formulas; Excel is apparently converting them behind the scenes.  The same spreadsheet imported into PHPSpreadsheet was outputting a #VALUE! error.

Added a function, modeled on the `convertToNumberIfFraction()` function, to detect if a string contains a percentage, and if so, convert it to a numeric.  If the last character of the string is a % sign, the function will string the % sign and divide the value by 100, overwriting the original param with the new value.

New helper function was added to `StringHelper.php` along the original faction helper.