Correct Some DocBlocks in Writer/CSV (#3225)
* Correct Some DocBlocks in Writer/CSV

The parameters for some methods had default values in PHPExcel, but the defaults were removed in PhpSpreadsheet. However, the doc-blocks continue to report a default. I also moved some of the parameter and return type declarations from doc-block to method signature. No other executable code was changed.

This isn't much of a change. I'm just pushing it now to see the effects of changes pushed earlier today (Php8.1 for Scrutinizer, no fake Php7.3 step).

* Comment Change

Scrutinizer was absent from a couple of merges today. Checking to see if it shows up here.